dynamic_sql_query.py
import mysql.connector
from config import DB_CONFIG

def execute_dynamic_query(query, params):
    connection = mysql.connector.connect(**DB_CONFIG)
    cursor = connection.cursor(dictionary=True)
    cursor.execute(query, params)
    result = cursor.fetchall()
    cursor.close()
    connection.close()
    return result

